Change shift to also shift truncated ID by default
This fixes an issue that arose from a missing shift(, True)
in AveragineCache. The option to not shift the truncated list
originally existed because I intended for the manipulation operations
to be efficient independent of order/intention to use all facets.
This is better served by opting out of the second shift rather than
explicitly asking for it.