Set default subgraph solver to "top"
The "top" solver tends to return the correct answer among
overlapped isotopic patterns. The "disjoint" solver is still
available. The "iterative" solver should not be used at this
time, as the algorithm it uses is not reliable for large
graphs.

Because the "top" solver is slower than "disjoint", I've tried
to introduce more C implementations for the slower code paths.