Feature: Allow ScanDataSources to provide their own peak picker
In order to test just how different the RawFileReader peak picker was,
I added a new branch to Scan.pick_peaks that dispatches off
fit_type="vendor". When triggered, this code tries to call
ScanDataSource._pick_peaks_vendor, which should return a PeakIndex. A
default implementation was added to ScanDataSource that raises a
NotImplementedError. When NotImplementedError is raised in
Scan.pick_peaks, it defaults back to ms_peak_picker.